+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release Date: ?
 
  - updated Debug Bar style and added new method for toggling it
  - fixed bug with Peak\Common\ExceptionLogger
+ - added new Peak\Common\Traits\UpdateToCamelCase
  
 VERSION 2.26.0
 --------------